+++ src/js/search-films.js
@@ -1,9 +1,7 @@
 import { Notify } from 'notiflix';
 import { searchFilms } from './api';
 import { createFilmItemMarkup } from './main-initial-films';
-import Pagination from 'tui-pagination';
-import 'tui-pagination/dist/tui-pagination.css';
-import { smoothScroll } from './scroll-to-top';
+import { pagination } from './pagination';
 
 const formSearchFilmsRef = document.querySelector('#search-form');
 const filmsList = document.querySelector('.films__list');
@@ -13,7 +11,7 @@ formSearchFilmsRef.addEventListener('submit', onSubmitFetchMovies);
 
 function onSubmitFetchMovies(e) {
   e.preventDefault();
-  filmsList.innerHTML = "";
+  filmsList.innerHTML = '';
   const keyword = e.currentTarget.searchQuery.value.trim();
   if (!keyword) {
     Notify.failure('Please enter something');
@@ -29,33 +27,13 @@ async function fetchFilmsByQuery(keyword, currentPage) {
     const {
       data: { results: filmsArray, total_results },
     } = await searchFilms(keyword, currentPage);
-
-    const paginationSearchFilms = new Pagination('pagination', {
-      totalItems: total_results,
-      itemsPerPage: filmsArray.length,
-      visiblePages: 5
-    });
-    paginationSearchFilms.on('afterMove', async (event) => {
-      const currentPage = event.page;
-      try {
-        spinner.hidden = false;
-        const {
-          data: { results: filmsArray },
-        } = await searchFilms(keyword, currentPage);
-        const items = createFilmItemMarkup(filmsArray);
-        filmsList.innerHTML = items;
-        smoothScroll();
-      } catch (error) {
-        Notify.failure(error.message);
-      } finally {
-        spinner.hidden = true;
-      }
-    });
-
     if (!filmsArray.length) {
       Notify.failure('Films not found');
       return;
     }
+
+    pagination(total_results, filmsArray, searchFilms, keyword);
+
     const items = createFilmItemMarkup(filmsArray);
     filmsList.innerHTML = items;
     spinner.hidden = true;
